HYD-5265 Update to Lustre build version 2.7.11.1
This lustre release introduces el7 support in parallel to
el6 support, so this patch adds the el7 packages as well
as updating the el6 packages. This requires some changes to
the makefiles to allow us to include packages from the
different distros simultaneously.

Update to the Lustre version 2.7.11.1 by changing
the following build parameters.

  JENKINS_JOB_NAME       = lustre-b_ieel3_0
  JENKINS_BUILD_NUMBER   = 40

The following chroma-dependencies are effected:
  lustre-{client,server}-{el6,el7,sles11}
  zfs

The above Lustre build no longer includes building the
xeon phi packages, so there are none we can include.
Doing a nice cleanup of the bundle and repo on upgrade
for existing releases will be handled in a future patch.

Also fixes a few minor issues to account for changes in
behavior with the new version.
  - Add chrome-dependencies/include/Makefile.distro
  - Update chroma-dependencies/include/Makefile.lustre_build_meta
  - Update the Makefiles for the following chroma-dependencies:
    zfs, lustre-client, lustre, robinhood, anyjson, el-extra, epel,
    nginx, robinhood
  - Remove chroma-dependencies/xeon-phi-client/*
  - Remove chroma-bundles/xeon-phi-client/*
  - Fixed a typo in the ZFS Makefile
  - Need an explicit 'modprobe zfs' after install now
  - Change in behavior of the mkdir mdt stat to count
    the lfs mkdir

Change-Id: I1df66f95fbde89cd745fbc36107f44c1941bdf2f
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/17811
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>