HYD-5358 Remove retired xeon-phi-client repo on upgrade
We have removed the xeon phi client packages and bundle from
EE. We should be polite and clean up the old xeon-phi-client
repo and protect our users from using the old release's repo
on accident.

We do this by adding to _install_or_upgrade_bundles, which is
what currently handles processing changes to bundles/repos upon
upgrade. This change deletes the bundle and leaves the old repo
in the BUNDLE_BACKUP location in case someone wants to retrieve
it.

  - add retired repo removal to _install_or_upgrade_bundles in
    chroma-bundles/install
  - add a test to the installation and upgrade test suite that
    checks that retired repos no longer appear after upgrade.

Change-Id: Idde7270fd339de3ceb5bf4fae3f59a9c5f3f013b
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/20628
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>