HYD-5362 Add function to agent to create the yum repo file.
At present the yum repo file is created at installation and cannot be updated
after that point. We have a need to change, add or delete repos from the file
and so a method on the agent to create the file is needed. This method can then
be called from the manager.

The method is be.

 create_yum_repo_file(filename, file_contents)
   filename      : is a name of the file that resides in /etc/yum.repos.d
   file_contents : is a string containing the contents of the file to write, an
                  empty string "" signifies the file should be deleted.

Durring the upgrade process this function is called to recreate the yum
repo. This requires a new step in the UpgradeJob call UpdateYumFileStep the
step will take the two parameters passed - the calculations should be done in
the Job not the Step.

 - Add create_yum_repo_file to agent_updates.py
 - Add UpdateYumFileStep and make UpdateJob use this new step. Update Job
   creates a repo file of all the manager repos which is a like for like
   behaviour of todays code. However in the future we could easily create
   references only to the repos that the target host requires.

Change-Id: Ib594f0807c953d2a2e55c5af776af0b1a5b36f11
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18179
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>