HYD-5366 Fix incompatibility between greenlet & gcc 4.8
greenlet (dependency of gevent) throws exceptions when
built with gcc 4.8 (default version on el7) as per:
http://stackoverflow.com/questions/15924353/gevent-redis-py-systemerror-null-result-without-error-in-pyobject-call

gevent is required to run at least one of the chroma
services orchestrated by supervisord when running
the manager.

The chosen solution is to install without optimisation
by supplying -O0 in CFLAGS during the python installation.

  - Add spec.ed for greenlet dependency to modify CFLAGS
    used in 'python setup.py install'
  - Update greenlet Makefile with note documenting change

Change-Id: Ia1839c7f2dac9579be3e7145ef55469b9142349c
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/18180
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>