HYD-5426 Fix for the create_ha differences between el6 and el7
Simple fix so that el6 and el7 clear ha commands are different
and selectable.

Also has the nice effect of tidying up the way the scripts are stored.

Be nice to have a better selection method, but even nicer to not need
the script!

It is also the case that the scripts are now so short and simliar that
having a single script might well be possible. But history tells me
that some fix might be required and if so it is better if that is
simple and issolated to the target. Hopefully the script can be
removed completely with all the actions being coded.

 - Add clear_ha_elX.sh scripts
 - Select script depending on the test distribution.

Change-Id: I0ea4a8f6b0a9e45028ac3b8881433b036d1e3b4a
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18189
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>