HYD-5381 Alert/Event migrations not complete
Alerts have been collapsed into a single table and  the additional
variables required for some old tables are now stored as json.

However the migration required to upgrade from < 3.0 to 3.0 was never
completed.

This patch completes that migration. For compatibility issues migrations
should not make use of the models in the application. Imagine the '3.0'
migration running when the app is upgrading to 4.0. The models will be those
of 4.0 not 3.0.

It is also the cases the storage resource (plugins) had some new attributes
which the current IML is not able to deal with - having no default for new
fields in old records. This is resolved by adding a optional default
attribute to storage records.

An additional pain is that the implementation of StepResult uses PickleField
which encodes a data in a way that means you cannot move a class between
modules or rename a class or do anything to a class and you cannot tell
that errors until runtime - so not only do we have a rubbish rushed
opaque previous implementaiton but we have to tests to know when it is
broken.  Anyway the solution is to pick apart one of the PickleFields (we
only have 2 - yes we imported a whole library to make two fields slightly
easier than they could have been - and actually the solution doesn't make
sense because we calculate something in real time that we should store
as a historical record)

 - Convert events to alerts because the notion of events disappeared
 - Convert the extra no standard extra alert fields into json
 - Add runtime tests to ensure the conversion is complete.
 - Rename, do not delete, the old tables so we can find them if it doesn't
   work in the field.
 - Enable some asserts that were commented out.
 - Rename one of the two 0025 migrations to 0026
 - Add default attribute to storage resource attribute. default can be
   a value or a callable.
 - Add simple tests to test the default attribute and the optional attribute
   which did not have any tests.

Change-Id: Ie16b338c8ab70be9b8c20b6e107762a9d2f25225
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18213
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>