HYD-5392 Update to Lustre build version 2.5.42.2
We need a new Lustre version to pull in a kernel update for the following:

A use-after-free flaw was found in the way the Linux kernel's key
management subsystem handled keyring object reference counting in certain
error path of the join_session_keyring() function. A local, unprivileged
user could use this flaw to escalate their privileges on the system.
(CVE-2016-0728, Important)

Update to the Lustre version 2.5.42.2 by changing the following build parameter:
  JENKINS_BUILD_NUMBER   = 230

The following chroma-dependencies are effected:
  lustre (el6.7)
  lustre-client-{el6.7,el7,sles11,sles12}
  zfs
  xeon-phi-client

  - Update chroma-dependencies/include/Makefile.lustre_build_meta

Change-Id: I334d6f4f7d8e565c0c02c0223e133f6c86690a94
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/18256
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>