HYD-5399 Add debug during agent deployment to logfile on target node
The setup/deployment/bootstrap script is remotely executed on the
target host after being retrieved via HTTP from the manager, this
curl request is executed on the target host by an agentRPC ssh call.

The agent deployment step can be hard to debug as the node executes a
python script remotely and doesn't communicate back to the manager
for a significant period of time.

Adding a remote debug log of this deployment step gives useful
information as to the cause of deployment failure and some
transparency into the bootstrapping process.

  - add identifiers to agent log formatters (daemon, console)
  - add logger to agent-bootstrap-script.template (bootstrap)
    which logs to chroma-agent.log on target
  - add logging for shell commands issued and bootstrap start/stop

Change-Id: I83b2f78869f92f1e74a6b092137944fcac412b30
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/18258
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>