HYD-5397 Make profile selection more user friendly
The GUI at present cannot easily tell if the validation
information in the host_profiles is valid or not. This
makes it difficult to present the best UX.

Thie change adds a profiles_valid flags to the host_profile
endpoint for each host to indicate if the validation
information is valid.  The profiles data is always present
in case the GUI times out on waiting for it to go valid.

A single test is added to make sure the profiles_valid=False
case occurs. A later change will test the opposite also
becomes True.

 - Add profiles_valid flag to endpoint.
 - Ensure that the flag is false immediately after host
   creation.

Change-Id: If87cba63ee3abc732594b47e34f27323d3920a5b
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18270
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>