HYD-5404 Build virtual envs on test nodes
The problem is that a couple of tests creates a virtualenv and
then uses pip to populate it, all on the job runner, which in
the case of the test-services test and SSI-simulator test is
the Jenkins master. This bakes in an assumption of what O/S
(more accurately, which Python version) that virtualenv will
be run with.

But then we pick up that virtualenv and copy it to the
provisioned chroma manager node and attempt to use it there,
which of course, in the case of an EL7 test, is an EL7 node,
which the virtualenv is no longer compatible with since EL7 has
Python 2.7 and EL6, where the virtualenv was created has Python 2.6.

The solution is to just create the virtualenvs where they need
to be.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I31817504c11dfbeb84cf646a79080abc524aeecb
Reviewed-on: http://review.whamcloud.com/18299
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>