HYD-5404 Build virtual envs on test nodes
The problem is that a couple of tests creates a virtualenv and
then uses pip to populate it, all on the job runner, which in
the case of the test-services test and SSI-simulator test is
the Jenkins master. This bakes in an assumption of what O/S
(more accurately, which Python version) that virtualenv will
be run with.

But then we pick up that virtualenv and copy it to the
provisioned chroma manager node and attempt to use it there,
which of course, in the case of an EL7 test, is an EL7 node,
which the virtualenv is no longer compatible with since EL7 has
Python 2.7 and EL6, where the virtualenv was created has Python 2.6.

The solution is to just create the virtualenvs where they need
to be.

Change-Id: Icd25cbd8d9d36a8436a6308b067971c43d4cbb49
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-on: http://review.whamcloud.com/20011
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>