HYD-5436 Lustre firewall rules not added for EL7
When running on EL7 the lustre firewall rules are not created, this is because
the rules are created within the spec file for the chroma-agent-management rpm
and so didn't benefit from the firewall updates.

The rules should be in the source and not in the chroma-agent-management spec
file for a number of reasons.

 The spec file is not a good place to have code.
 The chroma-agent-management rpm spec file is not the lustre rpm spec file and
 so should not make changes for the lustre rpm. It doesn't even have the lustre
 rpm as a dependency.

Add a agent command to create a firewall rule, add the corresponding remove
rule. Call the add rule for Lustre when creating a target.

 - Add OpenLustreFirewallStep to ConfigureTargetJob so that the firewall rule
   is created when a target is configured.
 - Add calls to open_firewall in OpenLustreFirewallStep
 - Add tests to test the open_firewall and close_firewall command
 - Firewall tests had to change because of HYD-5448, HYD-5448 means that we
   end up with multiple copies of the lustre rule in the tables.

Change-Id: I04546faa2b5471a7b05d79d3a99f76292738dc4c
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18361
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>