HYD-5434 NetworkManager clashes with IML changes.
When configuring the network via sysconfig scripts the
NetworkManager with EL7 the outcome requires more careful
interaction to stop NetworkManager from intefering with
the link.

Add an explicit unmange step to the nextwork configuration
so the when address updates take place we can be sure that
NetworkManager is not managing the device.

 - Update set address to unmange the device before setting
   the address
 - Add unmange_device function to node_admin.py
 - Edit the ifcfg-nic template to allow creating of sysconfig
   files without address/mask
 - Update tests for new functionality.

Change-Id: I412d8dab226af0e3402956b68ac473c73ebc8c46
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18366
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>