HYD-5431 No EL7 compatible profile for POSIX HSM Agent Node or Robinhood Policy Engine Server
When installing on EL7, the POSIX HSM Agent Node and Robinhood Policy Engine
Server profiles show as incompatible. Given the requirement for EL7 support for
worker nodes, this is a bug.

The compatibility display comes from validation rules in the profile which
are set for previous IML versions to only support 6.7. Removing this
restriction will show them as compatible for el6 and el7.

 - Remove validation rule from CT and RH profiles

Change-Id: Ice06b79b16c78674a5a525f1486e594baf925ca4
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18367
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>