HYD-5445 Power control monitor_daemon exception is raised when adding PDUs
The update of the fence agents in HYD-5107 erronously removed the requirement
for fence-agents on the manager for el7. This causes an execution error on
the manager when adding PDU's.

Durring this fix it also became apparent that the original HYD-5107 change was
in correct and rather than change 'status' to 'list-status' I should have
changes 'list' to 'list-status'. This change therefore undoes some work
that was done in that patch and corrects the errors.

Add the fence agent as a requirement into the manager spec file for el7 and
later, do the same for chroma-agent.spec as this has the same error.

 - Add requires: fence-agents to manager.spec file
 - Add requires: fence-agents to chroma-agent.spec file
 - Revert default_power_types.json back to original pre HYD-5107 state
 - Update boilerplate
 - Update power_control to change the correct field

Change-Id: I3d0650c9f81ade39b42767d827834a092a32667c
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18384
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>