HYD-5449 Allow None, True, False and json to be passed to the agent from the chroma-agent cli.
roma-agent allows any of the remote commands to be run.
chroma-agent mount-target --uuid=<uuid or target> however
some parameters need to be True/False/None or even json and
that can't be done from the CLI.

This patch makes that possible, it doesn't affect the rpc path only something
using the command line.

chroma-agent an only be run as root so anything that could be done
via this controls could be done directly.

 - Add a simple routine to turn known values like None,True,False
   into the computational value.
 - Also try loading each piece parameter to see if it is valid json

Change-Id: I62228deabddb50ad96d3fab351df657ea410aff9
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23175
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>