HYD-5256 Some legacy issues with blocking rpc calls
Status page Details buttons sometimes displays "Waiting" or
"Loading" and 30 seconds or more may elapse before dialog
appears. The root cause here is that the number of rpc's from
the Django GUI layer to the job schedular is limited meaning
that blocking long polling calls can eventually block all
calls.

The solution is to change the type of rpc used between the GUI
and the other processes (but primarily the job schedular) to
use a less lightweight implementation. It's not clear why we have
a lightweight and a 'heavyweight' implementation. All other processes
use 'heavyweight' version.  However this under 'extreme' circumstances
can now mean that the db connections can become overloaded on
the job schedular and so a throttling mechanism has been added (and
tests added) to limit the number of non-long-polling rpcs that
can be outstanding. We limit this to 5 which isn't many but as
Python isn't really multithreaded it is probably enough.

This patch disables the lightweight rpc's a follow on with remove
the now dead code completely. This will not be done here due to the
imminent release of 3.0.

  - Tweak to make authentication more efficient
  - Tweak to make debugging of long polling from a browser easier
  - Make sure connections are closed before blocking for a table
    change
  - Change to never use lightweight rpc's.
  - Add throttling to the rpc receiver.

Change-Id: I51218937c83b293fb39a075ad4084c3104daf58e
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18466
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>