HYD-5448 Firewall class doesn't detect existing rules properly
The add_rule is supposed to not duplicate a rule, but the duplication
test checks for a rule at the same position. However the index can
move and identical rules at a different index are not detected as
duplicates.

For our usage model the duplicate detection should ignore the index of
the existing rule and match any identical rules in the chain regardless
of order.

  - update FirewallControl class to implement new behaviour
  - update unit tests to verify new behaviour

Change-Id: I0e8f46c5f6af7316d4ccfd691a037b6a3c609095
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/18448
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>