HYD-5471 Unconfigured server should give option to pick profile
When the agent has deployed, but no server profile been selected
yet (for example, if the user chooses to navigate away while Deploy
Agent is running and then returns to complete the server setup), the
dropdown shows an option for "Setup monitored server" but no option
for any other profile type. This should behave in a way that puts all
of the profiles on even footing - either by instead bringing the user
to the dialog to select the server profile, or by giving an option in
the dropdown for each profile type. As it is it makes it look like
monitored only is the only path allowed forward.

This patch tidies up the drop down by enabling text on the verb change
of 'unconfigured' -> 'packages_installed'. It also refactors a slighty
modifies get_available_states for ManagedHost to make the packages_installed
state actionable. The refactor tidies what is still an awkard piece
of bootstrap code.

 - Update get_available_states to enable packages_installed
 - Add translatable verb text to InstallHostPackagesJob

Change-Id: I6f960bcdcbc7e6011c3d372a251644c75805e70a
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18497
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>