HYD-5495 Reverse dependencies missing for pacemaker.
The reverse dependency between a managed target and pacemaker is missing. This
means that if pacemaker is stopped the user will not be warned about the target
stopping and the target will not be stopped before pacemaker is stopped.

Strictly speaking of course you can stop pacemaker without the target
unmounting but for today at least we are probably most consistent if we say
that you can't have a filesystem running without pacemaker running.

Update the reverse dependencies and then add tests to ensure they are met.
Additionally add a test for lnet reverse dependencies.

 - Add reverse dependencies to corosync, corosync2 and target
 - Update test files to test these dependencies
 - Add test for lnet configuration

Change-Id: I76e84411e4e18ea76993c96fb481181e2c01b233
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18513
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>