HYD-5490 Status icon link not consistent between new and old ui pages
On the new UI pages, if you click on the Status icon when there are
active alerts, it will automatically apply a query to filter to the
currently active alerts. However, if you do this from an old ui page
(like filesystem detail), it just goes to the full list. New and old
pages not being obviously different to a user, this creates a sense of
confusion of what behavior to expect from the status icon, and makes you
think it's buggy, versus understanding the distinction.

This patch will ensure that the query string is added to the end of the
url if the health status is NOT good. If the status is good, the link
will remain at /ui/status. The link will be updated to use ng-href
instead of href and will call the "link" getter.

  - Added Link function to bigGreenButton.js. This getter will determine
    the link to the status page based on the health of the system.
  - Updated base.html such that the status link uses ng-href and calls
    the bigGreenButton controller's link getter function.

Change-Id: Ib08132253590e35eecdaf586fcaca4a60069c673
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/18517
Tested-by: Jenkins
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Tested-by: Joe Grund <joe.grund@intel.com>