HYD-5493 Exception printed on console during management install
During the installation of IML manager (chroma-manager) using the install
script, exceptions are displayed on the commandline to the user. These are
undesirable and should be written to the install log file instead.

"Stack traces just look so bad to a customer - it's the computer version of
being sick." - Chris, 18.02.16

The cause of this bug (identified using python2.7 on el7), is that subprocess
Popen doesn't catch exceptions within subprocess if stdout&stderr are not set to
PIPE. This is the case in the install script when passthrough option is set when
calling _run().  In this case the code that tries to catch exceptions when
subprocess is running chroma-config fails and the stack trace is displayed on
standard out.

These changes result in exceptions being correctly logged in the log file,
log.info statements being displayed to the user on the console, and raw_input
prompts being displayed to the user.

These issues were noticed in HYD-5470

  - exceptions are being output on stderr, in _run() passthrough mode,
    redirect stderr to the installation log file
  - due to bug (https://bugs.python.org/issue1927) where python redirects
    raw_input prompts to stderr unless GNU readlines module has been imported,
    import readlines
  - set the stream output in the logging streamhandler to sys.stdout so
    log.info messages from the service_config setup routine redirect to stdout
    rather than stderr, note different naming of StreamHandler parameter
    keyword (strm in python2.6, stream in python2.7)

Change-Id: I721a6c1ecf2b23202931af529923dfdf9c1e619c
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/18526
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>