HYD-5512 LNet incorrectly being stopped when removing non-managed host
When removing a monitored host IML would incorrectly try and unconfigure
LNet this would involve stopping and unloading it. This would fail because
the management modules were not present but the failure would also prevent
the node from being removed.

The change from managed/working->removed state on the manager is a different
transition from monitored->removed and so this should be described by
different transitions

This patch also fixes up a related unintended change from 2.4 which is that
worker nodes are not placed being placed into the lnet up state.

 - Add RemoveManagedHostJob transition to deal with managed and worker
   ost removal
 - Change dependency for worker node from lnet configured to lnet up

Change-Id: If03b2cbc78b299d2c1e44514107005e756d48851
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18580
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>