HYD-5527 User Accounts: the Email Alert type checkboxes are missing
Open either the current users account by clicking on the user account name in
the upper right of the IML interface, or open Configuration -> Users and click
Edit to open a user's settings, Click the Email Notifications tab

Notice there are no checkboxes

The issue is caused by the fact the list is produced from AlertState's
decendents, as part of the Alert/Event rationalization all Alerts/Events
became decendents of AlertStateBase.

This change accounts for that and explicitly removes AlertState from
the list.

 - Update get_object_list in AlertTypeResource to return children of
   AlertStateBase with AlertState filtered.

Change-Id: I32eeba09d47b910b8052a6675e785688a5134722
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18584
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>