HYD-5465 Auto select test profiles based on the target
To enable testing on multiple target distributions the profile
selected for test needs to be based on the target host. Because
we have profile validation rules we can use these rules and the
profile attributes to auto select the profile.

This patch is a response to some feedback in the review for change
id I3f6a97f41ab376c9482c832b19160d61bf7cdce1. Due to the importance
of the original patch it was landed as a seperate patch.

 - Remove some duplicate code in get_best_host_profile
 - Add some doc strings to get_host_validations

Change-Id: If17a2d717a17f3f23fa4e71662d6541bbb78fd22
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18667
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>