HYD-5572 Tests for NTP Conf start the mock for open multiple times.
Tests for NTP Conf start the mock for open multiple times, this means that when
the clean up occurs at the end of the test the mock may or may be cleared. It
is pure chance, start it twice and it is 50/50 start it 4 times at it is 1:4 -
the odds aren't quite that simple but there is a chance element.

Remove the extra starts for mock open and add the call the the super class for
TestNTPConfig because the is good practice.

 - Remove additional starts for mock_start.
 - Add call to super class for TestNTPConfig.

Change-Id: I8edccc921fac58fffa9405187ecc93ba5890974d
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18696
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>