HYD-5406 Unit test for profile validations fails on el7
This subtle bug appears when test run with python2.7 and
not with python2.6. Test expects error string to be returned == 'zero
length field…' this error is produced from formatted string in common
evaluators.py safe_eval() which uses format() with '{}' string
substitution Syntax not valid in python2.6. because this is valid in
python2.7 it returns a different error and thus the exception occurs.

After the above safe_eval() will evaluate input expressions and return
'Unsupported type...' error with object address when it encounters an
unrecognised variable.

It is more readable if this is changed to returning ValueError when an
unrecognised variable is encountered with the name of missing variable.

If the expression contains an embedded callable, return TypeError
with the ast class name as this is not needed/supported currently.

  - Fix safe_eval() in chroma-common/lib/evaluator.py to use python
    compatible with both 2.6 and 2.7
  - adjust tests to expect new error messages/exceptions rather than
    string formatting error
  - test for instances of ast.Name and raise ValueError in safe_eval()
  - test for any other instances of ast objects and raise TypeError
  - modify test_evaluator.py to expect ValueError for unrecognised
    variable names and TypeError for unrecognised types e.g. callables
  - add __init__.py to chroma-manager/tests/unit/chroma_common
    to enable test runner to find test files

Change-Id: I4decbdf4c2be22aca928b55132c5d9a3d02eab59
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/18711
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>