HYD-5582 AlertEvents do not have an end time set
AlertEvents do not have an end time set. This is
because they do not use the register_event of other
methods but just call save.

Simple update of the end_event handling code fixes this.

As part of the debug of this ticket is became obvious
that end_event for HostContactAlert had been lost and
so this event was replaced from the b2_4 branch.

 - Add correct handling of end_event in alert.py
 - Replace end_event for HostContactAlert from b2_4

Change-Id: Ib10f729cfdf3c64c52a41067eaca1843c995f83f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18718
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>