HYD-5559 PCS Password is hardcoded
The pcs password is hardcoded on the agent rather than being
generated.  Because we don't actually need to do anything other
than configure pcs we can generate a password randomly configure
pcs and then forget all about it.

Update the AutoConfigureCorosyncStep to generate a random 20
character password and use this for configuration.

Insert an intermeadiate function to mask the password from the
console log that is used. This means no-one ever knows the pcs
user password. If this is an issue then a sysadmin can change
it to a known password.

Create some tests for the Step and create a test class to allow
easy mocking of agent_invoke with the ability to make sure what
was intended to be run was run.

Addtionally: Because I copied CommandCaptureTestCase I realised
that a bug existed in that with regards to testing things had
actually been run, so I fix up that test code to work properly.

 - Add IMLUnitTestMockedInvoke test class
 - Add random passwords to AutoConfigureCorosyncStep
 - Add mask of random password from logs
 - Add tests for the new parts of AutoConfigureCorosyncStep
 - Fix bug found in CommandCaptureTestCase
 - Fix tests that were exploiting (potentially) this bug

Change-Id: I02900c48c021fff15e3da5febcdb6d0ae96c9d6f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18752
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>