HYD-5559 RabbitMQ Password is Hardcoded
The rabbitmq password is hardcode in the source. We already
have an exemption for it being fixed because the port is
only locally accessible we should still remove the hardcoding
from the source.

The definition is actually already duplicated and so this change
just removes the in code hardcoding and makes use of the value
in settings.py which would at least allow a user to define their
own value. Although that would not be super straight forward
because they would have to edit the file during the install
process.

 - Remove password defintions from service_config and use the
   values from settings.py

Change-Id: I388c3304c1f7c6a255340b1c774dff6a853c4740
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18761
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>