HYD-5509 Monitor mode shows lnet as unconfigured
When installing monitor mode the state of lnet is not updated
until it changes. The reason is that the manager is built to only
disipate changes of data from plugins into the code. The initial
update from lnet is before the manager is ready and so ignored,
further updates are filtered until a change. This is a general
race condition that exists for a number of states.

The solution is to not allow the plugins to create a session until
the manager is in a ready state. This means until the manager is in
the managed, monitored or working state.

 - Add a property to ManagedHost called ready_for_plugin_session_data
   which is true once the host is 'ready'
 - Check the property of the receiving host before allowing a session
   to be started.

Change-Id: I02d792037d0362d99c3f658bf5bcb9a38449355d
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18762
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>