HYD-5591 Robinhood policy engine fails to install: no mysql-server
Installing a server with the Robinhood Policy Engine Server profile
fails on el7.2. In el7, mysql-server is no longer in the vendor
repositories. (It does have a substitution available, called mariadb.)

Update the requires part of the spec file for robinhood to require
mariadb-server rather than mysql-server for el7. Do not change the
build parts as these do not differ for mariadb.

 - Update robinhood.spec requires to require mariadb-server for el7
   The change requires that mariadb-server and mariadb-devel is
   required, mariadb-devel provides mysql_config which is required
   by the robinhood policy engine.

Change-Id: I922d2ff0fc9f3f5975fd7c19a5a116b60dd2d158
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18803
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>